+++ lib/flutter_dom_ui.dart
@@ -16,3 +16,5 @@ export 'seo_widget/seo_textfield.dart';
 export 'seo_widget/seo_singlechildscrollview.dart';
 export 'seo_widget/seo_bottomnavigationbar.dart';
 export 'seo_widget/seo_footer.dart';
+export 'seo_widget/seo_sizedbox.dart';
+export 'seo_widget/seo_constrainedbox.dart';