+++ lib/seo_widget/seo_sizedbox.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dom_ui/helper/seo_injectable.dart';
 import 'package:flutter_dom_ui/dom_injector.dart';
 
-/// A container widget that not only displays content in Flutter,
+/// A SizedBox widget that not only displays content in Flutter,
 /// but also injects an equivalent HTML `<div>` structure for SEO purposes
 /// when running on web.
 ///
@@ -12,54 +12,26 @@ import 'package:flutter_dom_ui/dom_injector.dart';
 /// semantic HTML to search engines.
 ///
 /// Implements [SeoInjectableLayout] so it can be part of an SEO DOM hierarchy.
-class SeoContainer extends StatelessWidget implements SeoInjectableLayout {
-  /// The child widget to display inside the container.
+class SeoSizedBox extends StatelessWidget implements SeoInjectableLayout {
+  /// The child widget to display inside the SizedBox.
   final Widget? child;
 
-  /// How to align the child within the container.
-  final AlignmentGeometry? alignment;
-
-  /// Inner padding between the container’s edge and its child.
-  final EdgeInsetsGeometry? padding;
-
-  /// The background color of the container.
-  final Color? color;
-
-  /// A decoration to paint behind the child.
-  final Decoration? decoration;
-
-  /// A decoration to paint in front of the child.
-  final Decoration? foregroundDecoration;
-
-  /// The width of the container.
+  /// The width of the SizedBox.
   final double? width;
 
-  /// The height of the container.
+  /// The height of the SizedBox.
   final double? height;
-
-  /// Additional constraints to apply to the container.
-  final BoxConstraints? constraints;
-
-  /// Outer margin around the container.
-  final EdgeInsetsGeometry? margin;
 
   /// Optional callback triggered when the link is tapped in Flutter.
   final void Function()? onTap;
 
-  /// Creates a [SeoContainer] that works like a normal [Container]
+  /// Creates a [SeoSizedBox] that works like a normal [SizedBox]
   /// but also outputs an HTML `<div>` for SEO.
-  const SeoContainer({
+  const SeoSizedBox({
     super.key,
     this.child,
-    this.alignment,
-    this.padding,
-    this.color,
-    this.decoration,
-    this.foregroundDecoration,
     this.width,
     this.height,
-    this.constraints,
-    this.margin,
     this.onTap,
   });
 
@@ -69,23 +41,12 @@ class SeoContainer extends StatelessWidget implements SeoInjectableLayout {
       cursor: SystemMouseCursors.click,
       child: GestureDetector(
         onTap: onTap,
-        child: Container(
-          alignment: alignment,
-          padding: padding,
-          color: color,
-          decoration: decoration,
-          foregroundDecoration: foregroundDecoration,
-          width: width,
-          height: height,
-          constraints: constraints,
-          margin: margin,
-          child: child,
-        ),
+        child: SizedBox(width: width, height: height, child: child),
       ),
     );
   }
 
-  /// Injects the HTML representation of this container into the [parent] element.
+  /// Injects the HTML representation of this SizedBox into the [parent] element.
   ///
   /// Only executes if running on Web (`kIsWeb == true`).
   /// Creates a `<div>` element, assigns it a unique ID, and
@@ -96,21 +57,21 @@ class SeoContainer extends StatelessWidget implements SeoInjectableLayout {
     final document = webWindow.document;
 
     // Create a div element for HTML structure
-    final divContainer = document.createElement('div') as WebHTMLDivElement;
-    divContainer.id = _generateRandomId();
-    divContainer.style.width = width.toString();
-    divContainer.style.height = height.toString();
+    final divSizedbox = document.createElement('div') as WebHTMLDivElement;
+    divSizedbox.id = _generateRandomId();
+    divSizedbox.style.width = width.toString();
+    divSizedbox.style.height = height.toString();
 
     // Recursively inject child if present
     if (child != null) {
-      _appendWidgetToContainer(child!, divContainer);
+      _appendWidgetToContainer(child!, divSizedbox);
     }
 
     // Append this container to the parent HTML element
-    parent.appendChild(divContainer);
+    parent.appendChild(divSizedbox);
   }
 
-  /// Recursively appends a widget’s HTML equivalent to the container.
+  /// Recursively appends a widget’s HTML equivalent to the SizedBox.
   ///
   /// - If the widget implements [SeoInjectable], calls `createHtmlElement()`.
   /// - If it implements [SeoInjectableLayout], calls `injectHtmlTo()`.
@@ -126,6 +87,6 @@ class SeoContainer extends StatelessWidget implements SeoInjectableLayout {
   String _generateRandomId() {
     const chars = 'abcdefghijklmnopqrstuvwxyz0123456789';
     final rand = Random();
-    return 'container_${List.generate(4, (_) => chars[rand.nextInt(chars.length)]).join()}';
+    return 'sizedbox${List.generate(4, (_) => chars[rand.nextInt(chars.length)]).join()}';
   }
 }